requirements-dev.txt
# Requirements for local development and testing
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
flask>=2.3.2
flask-cors>=4.0.0
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2
pytest>=7.4.0
pytest-cov>=4.1.0
